CXX: removed exception windowspill test
This test randomly fails in CI. Hence the
removal until further investigation.
The original issue the test was addressing
has been fixed in our gcc; it has also been
confirmed and fixed in the upstream gcc.
This test only crashes at the end of the test
function so it is reasonable to assume that
it should be safe removing it so far, ie., the
test doesn't reveal a problem with the original
fix but a different one.

Opening an Jira issue IDFCI-76 for the further
investigation